requirements.txt
termcolor==1.1.0
opencv-python==4.5.3
torchvision
matplotlib